Merubah NIM jadi long

+++ Proyek/Mahasiswa.java
@@ -2,15 +2,15 @@ package Proyek;
 
 public class Mahasiswa
 {
-    private int nim;
+    private long nim;
     private String nama;
 
-    public void setNim(int nim)
+    public void setNim(long nim)
     {
         this.nim = nim;
     }
 
-    public int getNim()
+    public long getNim()
     {
         return this.nim;
     }
